Rename sfc TheControl
