+++ src/components/TheControl.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="example">{{ msg }}</div>
+</template>
+
+<script>
+export default {
+  data() {
+    return {
+      msg: "Hello world!",
+    };
+  },
+};
+</script>
+
+<style>
+.example {
+  color: red;
+}
+</style>
